Use dependency inversion for NetworkUtils
This PR introduced dependency inversion for the network Utils.

Pros:
- Removes the need to find an instance of connectivity manager each time a method is called
- Removes the need to pass in a context on each method call
- Removes the context dependency on NetworkUtils

This also introduces the singleton pattern. I used the application class to do this as itself is a singleton. If I was working on the project full time then I would suggest using [dagger2](https://google.github.io/dagger/) as a dependency
injection framework.

Removed:
- the method`getNetworkInfo` as this was ony used internally.
- `isConnectedWifi` as this was only used my the test
- the network test as it was mostly testing robolectric itself

There is a small amount of logic that isn’t tested in this class but it is essentially a null check and a boolean. IMHO it’s justified so the project can move away from the poorly maintained Roboblectric